Refactor analysis for nesting builder support
