+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/builder/AbstractDataFlowAnalysisBuilder.java
@@ -6,10 +6,10 @@ import java.util.List;
 import org.apache.log4j.Logger;
 import org.palladiosimulator.dataflow.confidentiality.analysis.DataFlowConfidentialityAnalysis;
 
-public abstract class AbstractDataFlowAnalysisBuilder<T extends DataFlowConfidentialityAnalysis, D extends AnalysisBuilderData> {
-	protected final Logger logger = Logger.getLogger(DataFlowConfidentialityAnalysisBuilder.class);
+public abstract class AbstractDataFlowAnalysisBuilder<T extends DataFlowConfidentialityAnalysis, D extends AnalysisBuilderData, E extends AnalysisBuilderData> {
+	protected final Logger logger = Logger.getLogger(DataFlowAnalysisBuilder.class);
 	protected final D builderData;
-	protected final List<AbstractDataFlowAnalysisBuilder<?, ?>> builder;
+	protected final List<AbstractDataFlowAnalysisBuilder<?, ?, ?>> builder;
 
 	/**
 	 * Creates a new builder with the given builder data
@@ -20,7 +20,11 @@ public abstract class AbstractDataFlowAnalysisBuilder<T extends DataFlowConfiden
 		this.builder = new ArrayList<>();
 	}
 	
-	public abstract void copyBuilderData(AnalysisBuilderData builderData);
+	/**
+	 * Copies the given expected builder data into the builder
+	 * @param builderData Expected builder data
+	 */
+	public abstract void copyBuilderData(E builderData);
 	
 	/**
 	 * Check the builder data that is saved in the builder
@@ -36,8 +40,9 @@ public abstract class AbstractDataFlowAnalysisBuilder<T extends DataFlowConfiden
 	 */
 	public abstract T build();
 	
-	public <B extends AbstractDataFlowAnalysisBuilder<?, ?>> B registerBuilder(B builder) {
+	public <B extends AbstractDataFlowAnalysisBuilder<?, ?, D>> B registerBuilder(B builder) {
 		this.builder.add(builder);
+		builder.copyBuilderData(builderData);
 		return builder;
 	}
