+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/builder/DataFlowAnalysisBuilder.java
@@ -2,19 +2,29 @@ package org.palladiosimulator.dataflow.confidentiality.analysis.builder;
 
 import org.palladiosimulator.dataflow.confidentiality.analysis.DataFlowConfidentialityAnalysis;
 
-public class DataFlowAnalysisBuilder extends AbstractDataFlowAnalysisBuilder<DataFlowConfidentialityAnalysis, AnalysisBuilderData>{
+public class DataFlowAnalysisBuilder extends AbstractDataFlowAnalysisBuilder<DataFlowConfidentialityAnalysis, AnalysisBuilderData, AnalysisBuilderData>{
 
 	public DataFlowAnalysisBuilder() {
 		super(new AnalysisBuilderData());
 		super.builder.add(this);
 	}
+	
+	@Override
+	public void copyBuilderData(AnalysisBuilderData builderData) {
+		super.builderData.setModelProjectName(builderData.getModelProjectName());
+		super.builderData.setStandalone(builderData.isStandalone());
+	}	
 
 	@Override
-	public void checkBuilderData() {}
+	public void checkBuilderData() {
+		if (!this.builderData.isStandalone()) {
+			throw new IllegalStateException("Execution of the analysis is only supported in standalone mode");
+		}
+	}
 
 	@Override
 	public DataFlowConfidentialityAnalysis build() {
 		this.checkBuilderData();
 		throw new IllegalStateException("No current implementation supports pcm-less analysis");
-	}	
+	}
 }
